qBittorrent serde: change u32 to i64

+++ src/librpl/qbittorrent.rs
@@ -73,14 +73,14 @@ pub enum State {
 
 #[derive(Debug, Deserialize, Getters)]
 pub struct QbitTorrentInfo {
-    added_on: u32,
+    added_on: i64,
     amount_left: u64,
     auto_tmm: bool,
     category: String,
     completed: i64,
-    completion_on: u32,
+    completion_on: i64,
     dl_limit: i64,
-    dlspeed: u64,
+    dlspeed: i64,
     downloaded: i64,
     downloaded_session: i64,
     eta: i64,
